Fixed asset pipeline issues with patternfly
The JS assets needs to be required ONLY in the application.js file,
and then this file is loaded by `javascript_link_tag('application')`.

PatternFly needs to be included as a bower dependency if you want to
have its JS dependencies. Unfortunately there is no other way :(
I assume that angular will be used in this repo, so I changed the
bower dependecy to `angular-patternfly-sass` which also loads the
JS deps the same way as `patternfly-sass`.